items work and more balanced, stopwatch is here

+++ Assets/Scripts/GameMaster.cs
@@ -12,7 +12,7 @@ public class GameMaster : MonoBehaviour
 
     public int gainzMultiplier = 1;
 
-    public static int gainz = 0;
+    public static long gainz = 0;
 
     public TextMeshProUGUI gainzText;
     public TextMeshProUGUI multiplierText;
